+++ lang/en/translation.php
@@ -1,0 +1,24 @@
+<?php
+
+return [
+    'greeting_morning' => 'Good Morning',
+    'greeting_afternoon' => 'Good Afternoon',
+    'greeting_evening' => 'Good Evening',
+    'greeting_night' => 'Good Night',
+    'gregorian_suffix' => 'AD',
+    'hijri_suffix' => 'AH',
+    'hijri_months' => [
+        1 => 'Muharram',
+        2 => 'Safar',
+        3 => "Rabi' al-Awwal",
+        4 => "Rabi' al-Thani",
+        5 => 'Jumada al-Awwal',
+        6 => 'Jumada al-Thani',
+        7 => 'Rajab',
+        8 => "Sha'ban",
+        9 => 'Ramadan',
+        10 => 'Shawwal',
+        11 => "Dhu al-Qi'dah",
+        12 => 'Dhu al-Hijjah',
+    ],
+];