+++ lang/id/translation.php
@@ -1,0 +1,24 @@
+<?php
+
+return [
+    'greeting_morning' => 'Selamat Pagi',
+    'greeting_afternoon' => 'Selamat Siang',
+    'greeting_evening' => 'Selamat Sore',
+    'greeting_night' => 'Selamat Malam',
+    'gregorian_suffix' => 'M',
+    'hijri_suffix' => 'H',
+    'hijri_months' => [
+        1 => 'Muharram',
+        2 => 'Safar',
+        3 => 'Rabiul Awwal',
+        4 => 'Rabiul Akhir',
+        5 => 'Jumadil Awwal',
+        6 => 'Jumadil Akhir',
+        7 => 'Rajab',
+        8 => "Sya'ban",
+        9 => 'Ramadhan',
+        10 => 'Syawwal',
+        11 => 'Zulqaidah',
+        12 => 'Zulhijjah',
+    ],
+];